Extract ImageView assertions
