+++ library/src/main/java/com/schibsted/spain/barista/assertion/BaristaAssertions.kt
@@ -1,15 +1,10 @@
 package com.schibsted.spain.barista.assertion
 
-import android.support.annotation.DrawableRes
-import android.support.annotation.IdRes
 import android.support.test.espresso.Espresso.onView
 import android.support.test.espresso.NoActivityResumedException
 import android.support.test.espresso.action.ViewActions
-import android.support.test.espresso.assertion.ViewAssertions.matches
 import android.support.test.espresso.matcher.ViewMatchers.isRoot
-import android.support.test.espresso.matcher.ViewMatchers.withId
 import com.schibsted.spain.barista.internal.failurehandler.RethrowingFailureHandler
-import com.schibsted.spain.barista.internal.matcher.DrawableMatchers.withDrawable
 import org.junit.Assert.fail
 
 object BaristaAssertions {
@@ -28,9 +23,4 @@ object BaristaAssertions {
         }
     }
 
-
-    @JvmStatic
-    fun assertDrawable(@IdRes id: Int, @DrawableRes drawable: Int) {
-        onView(withId(id)).check(matches(withDrawable(drawable)))
-    }
 }
